chore: change submodule upstream

+++ control/connection_info.go
@@ -1,13 +1,13 @@
 package control
 
 type ConnectionInfo struct {
-    Src      string
-    Dst      string
-    Sniffed  string
-    Dscp     uint8
-    L4proto  string
-    Dialer   string
-    Outbound string
-    Mac      string
-    Pname    string
+	Src      string
+	Dst      string
+	Sniffed  string
+	Dscp     int
+	L4proto  string
+	Dialer   string
+	Outbound string
+	Mac      string
+	Pname    string
 }
